# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ColdChain EMS API
  version: 1.0.0
  description: |
    Smart Cold-Storage & Energy Management System for Bangladesh operators.

    **Authentication**
    - Obtain an access token via `POST /api/v1/auth/login`.
    - Send it as `Authorization: Bearer <token>` on every protected request.
    - Refresh with `POST /api/v1/auth/refresh` before it expires (default 15 min TTL).

    **Branch context**
    - All tenant-scoped endpoints require `X-Branch-Id: <branch_id>` header.
    - Omitting it is only allowed for tenant-wide endpoints (admin users).

    **Money**
    All monetary values are in **poisha** (minor units). 1 BDT = 100 poisha.

    **Pagination**
    List endpoints return standard Laravel pagination. Use `?per_page=N&page=P`.

  contact:
    name: ColdChain EMS API Support
    email: api@coldchain.example
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker
  - url: https://api.coldchain.example/v1
    description: Production

# ─────────────────────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable parameters ──────────────────────────────────────────────────
  parameters:
    XBranchId:
      name: X-Branch-Id
      in: header
      description: Active branch context for this request.
      required: false
      schema:
        type: integer
        example: 1
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100

  # ── Common schemas ────────────────────────────────────────────────────────
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:    { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: The quantity field is required. }
            details: { type: object, nullable: true }
            request_id: { type: string, example: "01HXYZ…" }
    PaginationMeta:
      type: object
      properties:
        current_page: { type: integer }
        last_page:    { type: integer }
        per_page:     { type: integer }
        total:        { type: integer }
        from:         { type: integer, nullable: true }
        to:           { type: integer, nullable: true }
    PaginationLinks:
      type: object
      properties:
        first: { type: string, format: uri }
        last:  { type: string, format: uri }
        prev:  { type: string, format: uri, nullable: true }
        next:  { type: string, format: uri, nullable: true }

    # ── Auth ─────────────────────────────────────────────────────────────
...
    BranchResource:
      type: object
      properties:
        id:              { type: integer }
        tenant_id:       { type: integer }
        code:            { type: string }
        name:            { type: string }
        status:          { type: string, enum: [active, inactive, under_maintenance] }
        address_line1:   { type: string, nullable: true }
        address_line2:   { type: string, nullable: true }
        city:            { type: string, nullable: true }
        district:        { type: string, nullable: true }
        division:        { type: string, nullable: true }
        postal_code:     { type: string, nullable: true }
        country:         { type: string, nullable: true, example: BD }
        latitude:        { type: number, nullable: true }
        longitude:       { type: number, nullable: true }
        phone:           { type: string, nullable: true }
        email:           { type: string, nullable: true }
        timezone:        { type: string, nullable: true }
        chambers_count:  { type: integer, description: Present on index/show, which both eager-load the count }
        created_at:      { type: string, format: date-time }
        updated_at:      { type: string, format: date-time }
    StoreBranchRequest:
      type: object
      required: [code, name]
      properties:
        code:            { type: string, maxLength: 64, description: Unique within the tenant }
        name:            { type: string, maxLength: 255 }
        status:          { type: string, enum: [active, inactive, under_maintenance] }
        address_line1:   { type: string }
        address_line2:   { type: string }
        city:            { type: string }
        district:        { type: string }
        division:        { type: string }
        postal_code:     { type: string }
        country:         { type: string, minLength: 2, maxLength: 2, example: BD }
        latitude:        { type: number, minimum: -90, maximum: 90 }
        longitude:       { type: number, minimum: -180, maximum: 180 }
        phone:           { type: string }
        email:           { type: string, format: email }
        timezone:        { type: string }

    # ── Chambers ─────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Unauthenticated — token missing, invalid, or expired.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but lacking the required permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found (or belongs to another tenant).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State-machine conflict (e.g. editing a non-draft invoice).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnprocessableEntity:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }

# ─────────────────────────────────────────────────────────────────────────────
# Global security (overridden on public auth endpoints)
# ─────────────────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /branches:
    get:
      tags: [Branches]
      summary: List branches (requires `branches.view`)
      parameters:
        - $ref: '#/components/parameters/XBranchId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - { name: status, in: query, schema: { type: string, enum: [active, inactive, under_maintenance] } }
        - { name: q, in: query, schema: { type: string }, description: Search name or code }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:  { type: array, items: { $ref: '#/components/schemas/BranchResource' } }
                  meta:  { $ref: '#/components/schemas/PaginationMeta' }
                  links: { $ref: '#/components/schemas/PaginationLinks' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Branches]
      summary: Create branch (requires `branches.create`)
      parameters:
        - $ref: '#/components/parameters/XBranchId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreBranchRequest' }
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/BranchResource' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /branches/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
      - $ref: '#/components/parameters/XBranchId'
    get:
      tags: [Branches]
      summary: Show branch (requires `branches.view`)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/BranchResource' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Branches]
      summary: Update branch (requires `branches.update`)
      description: |
        `tenant_id` is immutable — a branch cannot be moved between tenants.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreBranchRequest' }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/BranchResource' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      tags: [Branches]
      summary: Delete branch (requires `branches.delete`)
      description: Fails with 409 if the branch still has chambers.
      responses:
        '200': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ── Chambers ───────────────────────────────────────────────────────────
...